Route01
type: feature
- implement browserRouter
- implement route
- create page for landing page
- create page for search page
- install package react-router-dom
