+++ package.json
@@ -6,6 +6,7 @@
     "prop-types": "^15.6.2",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "2.1.1"
   },
   "scripts": {
