+++ src/App.css
@@ -53,7 +53,7 @@ body, .app {
   bottom: 25px;
 }
 
-.open-search button {
+.open-search {
   display: block;
   width: 50px;
   height: 50px;
